# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para gestionar camisetas, clientes, tallas y calcular precio final segun el cliente consultado.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "operationId": "getClientes",
                "parameters": [
                    {
                        "name": "categoria",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear cliente",
                "operationId": "createCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_comercial",
                                    "rut",
                                    "direccion",
                                    "categoria",
                                    "contacto_nombre",
                                    "contacto_email"
                                ],
                                "properties": {
                                    "nombre_comercial": {
                                        "type": "string",
                                        "example": "90minutos"
                                    },
                                    "rut": {
                                        "type": "string",
                                        "example": "76123456-7"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Apoquindo 1234, Las Condes"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Preferencial"
                                    },
                                    "contacto_nombre": {
                                        "type": "string",
                                        "example": "Carla Paredes"
                                    },
                                    "contacto_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "compras@90minutos.cl"
                                    },
                                    "porcentaje_oferta": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener cliente por ID",
                "operationId": "getCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente",
                "operationId": "updateCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "operationId": "deleteCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
        "/clientes/{id}/camisetas": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar camisetas de un cliente",
                "operationId": "getClienteCamisetas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
...
    },
...
}